requirements.txt
accelerate==0.20.3
diffusers==0.17.0
Pillow==9.5.0
torch==2.0.1
transformers==4.30.1
